Move gltf folder to loaders folder
