+++ src/three/loaders/GLTFExtensionLoader.js
@@ -1,7 +1,7 @@
 import { DefaultLoadingManager } from 'three';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 import { LoaderBase } from '../../base/loaders/LoaderBase.js';
-import { GLTFCesiumRTCExtension } from '../gltf/GLTFCesiumRTCExtension.js';
+import { GLTFCesiumRTCExtension } from './gltf/GLTFCesiumRTCExtension.js';
 
 export class GLTFExtensionLoader extends LoaderBase {
