+++ src/three/loaders/gltf/GLTFStructuralMetadataExtension.d.ts
