+++ src/three/loaders/gltf/GLTFStructuralMetadataExtension.js
