+++ src/three/loaders/gltf/metadata/classes/ClassProperty.js
