+++ src/three/loaders/gltf/metadata/classes/MeshFeatures.js
