+++ src/three/loaders/gltf/metadata/classes/PropertyAttributeAccessor.js
