+++ src/three/loaders/gltf/metadata/classes/PropertySetAccessor.js
