+++ src/three/loaders/gltf/metadata/classes/PropertyTableAccessor.js
