+++ src/three/loaders/gltf/metadata/classes/PropertyTextureAccessor.js
