+++ src/three/loaders/gltf/metadata/classes/StructuralMetadata.js
