+++ src/three/loaders/gltf/metadata/utilities/ClassPropertyHelpers.js
