+++ src/three/loaders/gltf/metadata/utilities/TexCoordUtilities.js
