+++ src/three/loaders/gltf/metadata/utilities/TextureReadUtility.js
